refactor: session location
