+++ frontend/rust-lib/Cargo.toml
@@ -5,6 +5,7 @@ members = [
   "flowy-core",
   "dart-ffi",
   "flowy-user",
+  "flowy-user-wasm",
   "flowy-user-pub",
   "event-integration",
   "flowy-sqlite",
