+++ frontend/rust-lib/flowy-user-wasm/src/entities/mod.rs
@@ -1,1 +1,5 @@
 mod auth;
+mod user;
+
+pub use auth::*;
+pub use user::*;
