+++ frontend/rust-lib/flowy-user-wasm/src/event_handler.rs
@@ -1,11 +1,15 @@
-use flowy_error::FlowyError;
+use crate::entities::{OauthSignInPB, UserProfilePB};
+use crate::manager::UserManagerWASM;
+use flowy_error::{FlowyError, FlowyResult};
+use flowy_user_pub::entities::Authenticator;
 use lib_dispatch::prelude::{data_result_ok, AFPluginData, AFPluginState, DataResult};
-use std::sync::Weak;
+use lib_infra::box_any::BoxAny;
+use std::sync::{Arc, Weak};
 
 #[tracing::instrument(level = "debug", skip(data, manager), err)]
 pub async fn oauth_sign_in_handler(
   data: AFPluginData<OauthSignInPB>,
-  manager: AFPluginState<Weak<UserManager>>,
+  manager: AFPluginState<Weak<UserManagerWASM>>,
 ) -> DataResult<UserProfilePB, FlowyError> {
   let manager = upgrade_manager(manager)?;
   let params = data.into_inner();
@@ -14,4 +18,13 @@ pub async fn oauth_sign_in_handler(
     .sign_up(authenticator, BoxAny::new(params.map))
     .await?;
   data_result_ok(user_profile.into())
+}
+
+fn upgrade_manager(
+  manager: AFPluginState<Weak<UserManagerWASM>>,
+) -> FlowyResult<Arc<UserManagerWASM>> {
+  let manager = manager
+    .upgrade()
+    .ok_or(FlowyError::internal().with_context("The user session is already drop"))?;
+  Ok(manager)
 }
