+++ frontend/rust-lib/flowy-user/Cargo.toml
@@ -1,4 +1,3 @@
-workspace = { members = ["flowy-user-native", "user-native", "user-private", "user-wasm"] }
 [package]
 name = "flowy-user"
 version = "0.1.0"
