+++ frontend/rust-lib/flowy-user/src/entities/auth.rs
@@ -179,6 +179,26 @@ pub enum AuthenticatorPB {
   AppFlowyCloud = 2,
 }
 
+impl From<Authenticator> for AuthenticatorPB {
+  fn from(auth_type: Authenticator) -> Self {
+    match auth_type {
+      Authenticator::Supabase => AuthenticatorPB::Supabase,
+      Authenticator::Local => AuthenticatorPB::Local,
+      Authenticator::AppFlowyCloud => AuthenticatorPB::AppFlowyCloud,
+    }
+  }
+}
+
+impl From<AuthenticatorPB> for Authenticator {
+  fn from(pb: AuthenticatorPB) -> Self {
+    match pb {
+      AuthenticatorPB::Supabase => Authenticator::Supabase,
+      AuthenticatorPB::Local => Authenticator::Local,
+      AuthenticatorPB::AppFlowyCloud => Authenticator::AppFlowyCloud,
+    }
+  }
+}
+
 impl Default for AuthenticatorPB {
   fn default() -> Self {
     Self::Local
