+++ frontend/rust-lib/flowy-user/src/services/data_import/appflowy_data_import.rs
@@ -1,7 +1,7 @@
 use crate::migrations::session_migration::migrate_session_with_user_uuid;
 use crate::services::data_import::importer::load_collab_by_oid;
 use crate::services::db::UserDBPath;
-use crate::services::entities::{Session, UserPaths};
+use crate::services::entities::UserPaths;
 use crate::services::sqlite_sql::user_sql::select_user_profile;
 use crate::user_manager::manager_user_awareness::awareness_oid_from_user_uuid;
 use crate::user_manager::run_collab_data_migration;
@@ -28,6 +28,7 @@ use flowy_folder_pub::folder_builder::{ParentChildViews, ViewBuilder};
 use flowy_sqlite::kv::StorePreferences;
 use flowy_user_pub::cloud::{UserCloudService, UserCollabParams};
 use flowy_user_pub::entities::Authenticator;
+use flowy_user_pub::session::Session;
 use parking_lot::{Mutex, RwLock};
 use std::collections::{HashMap, HashSet};
 use std::ops::{Deref, DerefMut};
